build: ensure we get the latest version of urllib3
newer versions of urllib3 adopted the bad practise of resricting
the max python version, in this case <4. since observ is not
restricted in that sense (>=3.7), poetry looks for a version
of urllib3 that is not restricted either, and that's why it
kept resolving to 1.22. by explicitly stating in our dependency
specification (pyproject.toml) that we only want to use urllib3
when python <4, poetry is able to let go of this constraint
and selects the latest version.

this was important because older versions of urllib3 have known
security vulnerabilities, so we should to stop using it.