Fix compatibility with pygfx (#130)
* Fix compatibility with pygfx

This change prevents reactivity on custom dict subclasses and
instead will check that only real dicts are converted into
reactive proxies.
The problem is that pygfx's geometry is a dict subclass and
turning it into a reactive proxy will stop it from being recognized
as a Geometry type, so pygfx will not accept it when you try to
set the geometry attribute on a Mesh.
I could check for pygfx specifically, but it seems to me that
more problems are to be expected in the future if observ tries
to proxy custom dict subclasses.
I'm thinking to update list and set proxy as well.

* Fix linting by formatting

* Remove uv.lock